requirements/base.txt
Django=1.11.1
psycopg2=2.7.1
djangorestframework=3.6.2